+++ packages/contracts/src/loot/RandomEquipmentSubsystem.sol
@@ -21,7 +21,7 @@ contract RandomEquipmentSubsystem is Subsystem {
   /// @param randomness used to randomly pick equipment prototype and affixes.
   /// @return lootEntity a new entity.
   function executeTyped(
-    uint256 ilvl,
+    uint32 ilvl,
     uint256 randomness
   ) public returns (uint256 lootEntity) {
     return abi.decode(
@@ -32,9 +32,9 @@ contract RandomEquipmentSubsystem is Subsystem {
 
   function _execute(bytes memory args) internal override returns (bytes memory) {
     (
-      uint256 ilvl,
+      uint32 ilvl,
       uint256 randomness
-    ) = abi.decode(args, (uint256, uint256));
+    ) = abi.decode(args, (uint32, uint256));
 
     // pick equipment prototype (it's the targetEntity when getting affix availability)
     uint256 equipmentProtoEntity = LibLootEquipment.pickEquipmentPrototype(ilvl, randomness);
